## 4.6.0.0
* Feature Additions
  * User can configure toolbar to left or right side of screen
  * Updated self-marker to allow color customizations via Settings > Display Preferences > Color/Tinting > My Location Color/Size
  * End user bug reporting and feedback via Settings > Support > User Feedback. Reports are sent to the connected TAK Server.
  * Updated user preferences L&F
  * Drawing tools improvements
    * Added minimum safe distance (MSD) to shapes. MSD may be controlled via a shape's radial menu.
    * Manual parameters input for Rectangle and Ellipse shapes
    * End points for polyline may be manually edited via the details pane
    * Additional line styles for shapes now available
* Bug Fixes
  * Better manage device screen always on mode when navigating.
  * Added preference to Air Overlays plugin to disable Bluetooth/ADS-B scanning.
  * Resolved issue with rendering KMZs using Rubber Sheet
  * Resolved reported crash issue with screen lock (via play store).
  * Resolved issue where a user could not attach pictures from the camera to a marker
  * Resolved issue with Pathways not being displayed in 3D mode.
* Security
  * Disabled HTTP transfer of Data Packages

+++ takkernel/engine/src/main/cpp/renderer/core/GLMapViewDebug.cpp
@@ -67,7 +67,7 @@ void TAK::Engine::Renderer::Core::GLMapViewDebug_drawCameraRangeRing(const GLGlo
     xyz.reserve(180u*size);
 
     for(std::size_t i = 0u; i < 180u; i++) {
-        GeoPoint2 lla = GeoPoint2_pointAtDistance(GeoPoint2(cam.latitude, cam.longitude), i*2, radiusMeters, true);
+        GeoPoint2 lla = GeoPoint2_pointAtDistance(GeoPoint2(cam.latitude, cam.longitude), static_cast<double>(i*2), radiusMeters, true);
         TAK::Engine::Math::Point2<double> p;
         view.renderPass->scene.forward(&p, lla);
         xyz.push_back((float)p.x);
@@ -186,9 +186,9 @@ namespace
 
         atakmap::renderer::GLES20FixedPipeline::getInstance()->glEnableClientState(atakmap::renderer::GLES20FixedPipeline::CS_GL_VERTEX_ARRAY);
 
-        atakmap::renderer::GLES20FixedPipeline::getInstance()->glVertexPointer(size, GL_FLOAT, 0, pos);
+        atakmap::renderer::GLES20FixedPipeline::getInstance()->glVertexPointer(static_cast<int>(size), GL_FLOAT, 0, pos);
 
-        atakmap::renderer::GLES20FixedPipeline::getInstance()->glDrawArrays(mode, 0, count);
+        atakmap::renderer::GLES20FixedPipeline::getInstance()->glDrawArrays(mode, 0, static_cast<int>(count));
 
         atakmap::renderer::GLES20FixedPipeline::getInstance()->glDisableClientState(atakmap::renderer::GLES20FixedPipeline::CS_GL_VERTEX_ARRAY);
